requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
dnspython==2.7.0
gunicorn==23.0.0
MarkupSafe==3.0.2
mongoengine==0.29.1
packaging==25.0
PyJWT==2.10.1
pymongo==4.11.3
sqlparse==0.5.3
tzdata==2025.2
waitress==3.0.2
Werkzeug==3.1.3
whitenoise==6.9.0
